Add files via upload
Specifically, K-means tends to perform better when centroids are seeded in such a way that does not clump them together in space.
 Step 1: Choose one of your data points at random as an initial centroid. 
Step 2: Calculate D(x), the distance between your initial centroid and all other data points, x.
 Step 3: Choose your next centroid from the remaining data points with probability proportional to D(x).D(x)
 Step 4: Repeat until all centroids have been assigned. 
Note: D(x) should be updated as more centroids are added. It should be set to be the distance between a data point and the nearest centroid.